+++ Classes/Domain/Model/Event/EventTopic.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OliverKlee\Seminars\Domain\Model;
+namespace OliverKlee\Seminars\Domain\Model\Event;
 
 /**
  * This class represents a topic for an event. It can be referenced by multiple event dates.