+++ Classes/Domain/Model/Event/SingleEvent.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OliverKlee\Seminars\Domain\Model;
+namespace OliverKlee\Seminars\Domain\Model\Event;
 
 /**
  * This class represents a single event, e.g., an event has the topic and date all in one record.