+++ Tests/Unit/Domain/Repository/Event/EventRepositoryTest.php
@@ -2,15 +2,15 @@
 
 declare(strict_types=1);
 
-namespace OliverKlee\Seminars\Tests\Unit\Domain\Repository;
+namespace OliverKlee\Seminars\Tests\Unit\Domain\Repository\Event;
 
 use Nimut\TestingFramework\TestCase\UnitTestCase;
-use OliverKlee\Seminars\Domain\Repository\EventRepository;
+use OliverKlee\Seminars\Domain\Repository\Event\EventRepository;
 use TYPO3\CMS\Extbase\Object\ObjectManagerInterface;
 use TYPO3\CMS\Extbase\Persistence\Repository;
 
 /**
- * @covers \OliverKlee\Seminars\Domain\Repository\EventRepository
+ * @covers \OliverKlee\Seminars\Domain\Repository\Event\EventRepository
  */
 final class EventRepositoryTest extends UnitTestCase
 {